% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_likert.R
\name{as_prop}
\alias{as_prop}
\title{Utility functions to transform likert scale values to proportions & log-odds.}
\usage{
as_prop(x, n, range)
}
\arguments{
\item{x}{\verb{Numeric vector} Summed likert value to transform.}

\item{n}{\verb{Numeric scalar} How many items were summed in \code{x}.}

\item{range}{\verb{Numeric scalar} Number of possible answers in the likert scale.}
}
\description{
Utility functions to transform likert scale values to proportions & log-odds.
}
