% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting_functions.R
\name{cum_event_rate_chart}
\alias{cum_event_rate_chart}
\title{Generate a cumulative event rate chart}
\usage{
cum_event_rate_chart(act, pred, font_size = 14, bad = 0)
}
\arguments{
\item{act}{A numeric vector of actual 0s/1s}

\item{pred}{A numeric vector of predicted probabilities (or log odds)}

\item{font_size}{A number specifying the general font size to be used in all the charts}

\item{bad}{0/1 sepcifying what is \code{bad} (i.e. default)}
}
\description{
Function generates a chart to visualise cumulative event rates and associated approval rates
}
\details{
Cumulative event rates are generated assuming everyone above a certain output threshold is approved.
Approval rate is calculated as the population percentage associated with each threshold.
}
