% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting_functions.R
\name{gains_chart}
\alias{gains_chart}
\title{Generate a gains chart (not exported)}
\usage{
gains_chart(obj, font_size = 14)
}
\arguments{
\item{obj}{An object of type \code{ks} ()}

\item{font_size}{A number specifying the general font size to be used in all the charts}
}
\description{
Function generates a chart to visualise the gains table.
This function is the default plot method for objects created with the \code{gains_table()} function.
}
\details{

}
