% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_functions.R
\name{gains_table}
\alias{gains_table}
\title{Generate a Gains table}
\usage{
gains_table(act, pred, nBins = 10, prettify = T, type = "goods")
}
\arguments{
\item{act}{A numeric vector of actual observed classes}

\item{pred}{A numeric vector of predictions (probabilities/log-odds/scores)}

\item{nBins}{An integer specifying the number of bins to use (10 = deciles, 20 = semi deciles etc)}

\item{prettify}{A boolean specifying if the function should return a pretty table (huxtable will be used)}

\item{type}{A string specifying if event = goods or bads. Output will be inverted accordingly.}
}
\value{
A tibble if \code{prettify = T} else an object of type \code{ks}
}
\description{
Function generates a gains table that is typically used to assess the performance of scorecards.
The primary outputs to look at would be:

\enumerate{
\item Event rate trend
\item Value of the KS statistic
\item Capture rates
\item Cumulative Event Rates for threshold selection
}
}
\details{
The function relies on provided vectors for observed classes and predictions to create
a bin wise summary. Typically, this is done by dividing the population into deciles, or
10 separate bins. While the function supports any number of bins but going beyond 20 bins
is not recommended since individual computations like event rates become unreliable.

The KS statistic is the primary performance measure when using a gains table. The function
also computes the AUROC and the Gini coefficient as \code{2*AUROC - 1}.
}
\examples{
data("accepted_base")
accepted_base <- na.omit(accepted_base)

with(accepted_base, {
 bad_flag <- ifelse(loan_status == "Charged Off", 1, 0)
 mdl <- glm(factor(bad_flag) ~ revol_util + last_pymnt_amnt,
            family = "binomial")
 gains_table(bad_flag, predict(mdl), nBins = 10, type = "bads")
})
}
