% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_functions.R
\name{IV}
\alias{IV}
\title{Generate a diagnistic plot}
\usage{
IV()
}
\arguments{
\item{act}{A numeric vector of actual observed classes}

\item{pred}{A numeric vector of predictions (probabilities/log-odds/scores)}

\item{nBins}{An integer specifying the number of bins to use (10 = deciles, 20 = semi deciles etc)}

\item{prettify}{A boolean specifying if the function should return a pretty table (huxtable will be used)}

\item{type}{A string specifying if event = goods or bads. Output will be inverted accordingly.}
}
\value{
A tibble if \code{prettify = T} else an object of type \code{ks}
}
\description{
Function generates a set of diagnostic plots to help evaluate the performance of a predictive model. The following four
charts are shown:

\enumerate{
\item Gains chart
\item ROC Curve
\item Capture rates
\item KS Chart
}
}
\details{

}
\examples{
data("accepted_base")
accepted_base <- na.omit(accepted_base)

tab <- with(accepted_base, {
 bad_flag <- ifelse(loan_status == "Charged Off", 1, 0)
 mdl <- glm(factor(bad_flag) ~ revol_util + last_pymnt_amnt,
            family = "binomial")
 gains_table(bad_flag, predict(mdl), nBins = 10, type = "bads")
})
gains_chart(tab)
}
