% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{linear_scale}
\alias{linear_scale}
\title{Generate scores using a linear transformation}
\usage{
linear_scale(
  logodds,
  points_to_scale = 20,
  scale = 2,
  anchor = 700,
  odds_at_anchor = 0.1,
  plot = T
)
}
\arguments{
\item{logodds}{A numeric vector of LogOdds}

\item{points_to_scale}{An integer specifying the number of points(score points i.e.) which would trigger a 'x' times change in log odds}

\item{scale}{An integer specifying the factor why which \code{points_to_scale} will change log odds. (Ex: 2 would mean doubling of odds)}

\item{anchor}{An integer used to anchor the score output. See \code{details}}

\item{odds_at_anchor}{A numeric value specifying the log odds at the anchor point}

\item{plot}{A True/False value specifying if the function should plot a histogram of the generated scores}
}
\description{
Function to convert log odds to an easy to interpret numeric score.
}
\details{
Scores are generated as linear trnasformation of log odds such that \eqn{score = \alpha + \beta \times LogOdds}.
Where,
\deqn{\alpha = PDO/log(scale)}
\deqn{\beta - Anchor - \alpha * OddsAtAnchor}
}
