% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting_functions.R
\name{precision_recall_chart}
\alias{precision_recall_chart}
\title{Generate a precision-recall chart (not exported)}
\usage{
precision_recall_chart(act, pred, font_size = 14)
}
\arguments{
\item{act}{A numeric vector of actual 0s/1s}

\item{pred}{A numeric vector of predicted probabilities (or log odds)}

\item{font_size}{A number specifying the general font size to be used in all the charts}
}
\description{
Function generates a chart to visualise precision and recall values across cutoffs
}
\details{

}
