% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting_functions.R
\name{roc_chart}
\alias{roc_chart}
\title{Generate a ROC Curve}
\usage{
roc_chart(act, pred, font_size = 14)
}
\arguments{
\item{act}{A numeric vector of actual 0s/1s}

\item{pred}{A numeric vector of predicted probabilities (or log odds)}

\item{font_size}{A number specifying the general font size to be used in all the charts}
}
\description{
Function generates an ROC Curve.
}
\details{

}
\examples{
data("accepted_base")
accepted_base[is.na(accepted_base)] <- -999

with(accepted_base, {
 bad_flag <- ifelse(loan_status == "Charged Off", 1, 0)
 mdl <- glm(factor(bad_flag) ~ revol_util + last_pymnt_amnt,
            family = "binomial")
 roc_chart(bad_flag, predict(mdl))
})
}
