% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charting_functions.R
\name{sensitivity_specificity_chart}
\alias{sensitivity_specificity_chart}
\title{Generate a sensitivity-specificity chart (not exported)}
\usage{
sensitivity_specificity_chart(act, pred, font_size = 14)
}
\arguments{
\item{act}{A numeric vector of actual 0s/1s}

\item{pred}{A numeric vector of predicted probabilities (or log odds)}

\item{font_size}{A number specifying the general font size to be used in all the charts}
}
\description{
Function generates a chart to visualise sensitivity and specificity values across cutoffs
}
\details{

}
