% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_functions.R
\name{diag_charts}
\alias{diag_charts}
\title{Generate a diagnostic plot}
\usage{
diag_charts(act, pred, font_size = 14)
}
\arguments{
\item{act}{A numeric vector of actual observed classes}

\item{pred}{A numeric vector of predictions (probabilities/log-odds/scores)}

\item{font_size}{A number specifying the general font size to be used in all the charts}
}
\description{
Function generates a set of diagnostic plots to help evaluate the performance of a predictive model.
}
\details{

}
\examples{
data("accepted_base")
accepted_base[is.na(accepted_base)] <- -999

with(accepted_base, {
bad_flag <- ifelse(loan_status == "Charged Off", 1, 0)
mdl <- glm(factor(bad_flag) ~ revol_util + last_pymnt_amnt,
          family = "binomial")
diag_charts(bad_flag, predict(mdl, type = "response"))
})
}
